Merged PR 16: Tutorial creation positive and negative example for link prediction
<!--- Provide a general summary of your changes in the Title above -->
## Description

<!--- Describe your PR Changes here-->

Due the importance of the creation of positive and negative example for training link prediction models and with the purpose of maintaining the notebooks as short and clean as possible, I have created a stand-alone notebook with the description of the necessary steps for creating positive and negative example in link prediction models. The tutorial uses the FinCEN graph.
### PR Checklist (if applicable)

- [x] Your code builds clean without any errors or warnings
- [x] The documentation has been updated to account for the changes made
- [x] Code and markdown is formatted

Related work items: #44