Merged PR 14: Link prediction
<!--- Provide a general summary of your changes in the Title above -->
## Description

<!--- Describe your PR Changes here-->

A simplified version of the link prediction demo is now created.  It describes the building blocks needed for performing Link Prediction with GNN.   two different Neural Network layers are presented. The FinCEN dataset is used for example for training. Results are provided with performance metrics and graphs

### PR Checklist (if applicable)

- [x] Your code builds clean without any errors or warnings
- [ ] The documentation has been updated to account for the changes made
- [x] Code and markdown is formatted

Related work items: #44