Merged PR 51: update dataset documentation
<!--- Provide a general summary of your changes in the title above -->
## Description

<!--- Describe your PR Changes here-->

A new dataset has been added to _public/tools_ container hosting a the list of al countries of the world with and their correspondent ISO codes, geo locations, etch. It is used for plotting the edges representing the mistenkelige transactions between banks across the world as results from the GNN results.

### PR Checklist (if applicable)

- [ ] Your code runs without errors
- [x] The documentation has been updated to account for the changes made
- [ ] Code and markdown is formatted
- [ ] The code contains no credentials, secrets or sensitive information

### Infrastructure checklist (if applicable)

- [ ] Infrastructure files are formatted and validated
- [ ] Pipeline runs without errors

### Demos details (if applicable)

**Main audience:** <!--- Describe which audience the demo is intended for -->